Refactor: Rename sentiment to feedback
